chore: 소소한 수정
parsorTest
- 테스트 문자열 변경

LottoRank, Statistics
- 상금 int -> long 타입으로 변경

+++ src/main/java/lotto/domain/Statistics.java
@@ -3,12 +3,12 @@ package lotto.domain;
 import java.util.List;
 
 public class Statistics {
-    private int FIRST;
-    private int SECOND;
-    private int THIRD;
-    private int FOURTH;
-    private int FIFTH;
-    private int purchase;
+    private long FIRST;
+    private long SECOND;
+    private long THIRD;
+    private long FOURTH;
+    private long FIFTH;
+    private long purchase;
 
     public Statistics(){
         FIRST = 0;
@@ -33,7 +33,7 @@ public class Statistics {
         purchase += 1000;
     }
 
-    public int getByRank(LottoRank rank) {
+    public long getByRank(LottoRank rank) {
         if(rank == LottoRank.FIRST) return FIRST;
         if(rank == LottoRank.SECOND) return SECOND;
         if(rank == LottoRank.THIRD) return THIRD;
@@ -42,7 +42,7 @@ public class Statistics {
         return 0;
     }
 
-    public int getPrizeAmount() {
+    public long getPrizeAmount() {
         return FIRST * LottoRank.FIRST.PRIZE
                 + SECOND * LottoRank.SECOND.PRIZE
                 + THIRD * LottoRank.THIRD.PRIZE
